+++ app/src/main/java/org/simple/clinic/monthlyReports/questionnaire/entry/epoxy/model/InputGroupEpoxyModel.kt
@@ -1,0 +1,46 @@
+package org.simple.clinic.monthlyReports.questionnaire.entry.epoxy.model
+
+import android.annotation.SuppressLint
+import android.view.View
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.airbnb.epoxy.EpoxyAttribute
+import com.airbnb.epoxy.EpoxyHolder
+import com.airbnb.epoxy.EpoxyModelClass
+import com.airbnb.epoxy.EpoxyModelWithHolder
+import org.simple.clinic.R
+import org.simple.clinic.monthlyReports.questionnaire.component.InputViewGroupComponentData
+
+@SuppressLint("NonConstantResourceId")
+@EpoxyModelClass(layout = R.layout.view_questionnaire_input_group)
+abstract class InputGroupEpoxyModel : EpoxyModelWithHolder<InputGroupEpoxyModel.Holder>() {
+
+  @EpoxyAttribute
+  lateinit var inputViewGroupComponentData: InputViewGroupComponentData
+
+  override fun bind(holder: Holder) {
+    val items = inputViewGroupComponentData.children
+    if (items != null) {
+      setLayoutManager(holder.recyclerView, items.count())
+
+      val inputFieldAdapter = InputFieldAdapter()
+      holder.recyclerView.adapter = inputFieldAdapter
+      inputFieldAdapter.submitList(items)
+    }
+  }
+
+  private fun setLayoutManager(recyclerView: RecyclerView, itemCount: Int) {
+    recyclerView.layoutManager = GridLayoutManager(
+        recyclerView.context,
+        if (itemCount > 1) 2 else 1
+    )
+  }
+
+  class Holder : EpoxyHolder() {
+    lateinit var recyclerView: RecyclerView
+
+    override fun bindView(itemView: View) {
+      recyclerView = itemView.findViewById(R.id.inputFieldRecyclerView)
+    }
+  }
+}